Fixed S3BotoStorage performance problem calling modified_time()
For running collectstatic on 10000 static files, it's the difference between <1s and minutes.

S3BotoStorage.modified_time was calling get() on entry and performing a remote call to populate the default (which only matters if get() fails to find the key)

I merely added a line to check if entry is None, then make the remote call.